pyproject.toml
[build-system]
requires = ["setuptools==75.6.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "apcos"
version = "1.0.0rc1"
description = "Adaptive Personal Cognitive Operating System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML==6.0.2",
  "psutil==6.1.1",
]

[project.optional-dependencies]
dev = [
  "pytest==9.0.2",
  "hypothesis==6.151.9",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

